+++ packages/php/examples/example2_client.php
@@ -14,7 +14,7 @@ $transport = new DoctrineIpTransport($connection, uniqid('transport_', true));
 $client = new Client($transport, $transport);
 
 $ip = long2ip(random_int(ip2long('10.0.0.0'), ip2long('10.255.255.255')));
-for ($i = 0; $i < 3; ++$i ) {
+for ($i = 0; $i < 3; ++$i) {
     $data = new ArrayObject([
         'client' => $ip,
         'id' => $i,